# api/swagger/swagger.json
{
  "info": {
    "title": "Location Service",
    "description": "This service is responsible for delivering messages to / from applications",
    "version": "1.0.0"
  },
  "swagger": "2.0",
  "paths": {
...
    "/accounts": {
      "post": {
        "description": "Save Inflight Account",
        "operationId": "saveAccount",
        "tags": [
          "account"
        ],
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "account",
            "description": "InflightAccount",
            "type": "object",
            "schema": {
              "$ref": "#/definitions/InflightAccount"
            },
            "in": "body",
            "required": true
          },
          {
            "name": "X-Tenant-Id",
            "description": "Tenant Id",
            "type": "string",
            "in": "header",
            "required": false
          },
          {
            "name": "x-fast-pass",
            "description": "Bypass Auth",
            "type": "boolean",
            "in": "header",
            "require": false
          }
        ],
        "responses": {
          "201": {
            "description": "InflightAccount",
            "type": "object",
            "schema": {
              "$ref": "#/definitions/InflightAccount"
            }
          },
          "409": {
            "description": "Conflict",
            "type": "object",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "500": {
            "description": "Internal Error",
            "type": "object",
            "schema": null
          }
        }
      },
      "get": {
        "description": "Page Accounts",
        "operationId": "page",
        "tags": [
          "account"
        ],
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "query",
            "description": "Query String - default '{}'",
            "type": "string",
            "in": "query",
            "required": false
          },
          {
            "name": "limit",
            "description": "Limit - default 10",
            "in": "query",
            "required": false
          },
          {
            "name": "offset",
            "description": "Offset - default 0",
            "in": "query",
            "required": false
          },
          {
            "name": "X-Tenant-Id",
            "description": "Tenant Id",
            "type": "string",
            "in": "header",
            "required": false
          },
          {
            "name": "x-fast-pass",
            "description": "Bypass Auth",
            "type": "boolean",
            "in": "header",
            "require": false
          }
        ],
        "responses": {
          "200": {
            "description": "PageResult",
            "type": "object",
            "schema": {
              "$ref": "#/definitions/PageResult"
            }
          },
          "500": {
            "description": "Internal Error",
            "type": "object",
            "schema": null
          }
        }
      }
    },
    "/accounts/{id}": {
      "put": {
        "description": "Update Inflight Account",
        "operationId": "updateAccount",
        "tags": [
          "account"
        ],
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "description": "Id of InflightAccount to update",
            "type": "string",
            "in": "path",
            "required": true
          },
          {
            "name": "account",
            "description": "InflightAccount",
            "type": "object",
            "schema": {
              "$ref": "#/definitions/InflightAccount"
            },
            "in": "body",
            "required": true
          },
          {
            "name": "X-Tenant-Id",
            "description": "Tenant Id",
            "type": "string",
            "in": "header",
            "required": false
          },
          {
            "name": "x-fast-pass",
            "description": "Bypass Auth",
            "type": "boolean",
            "in": "header",
            "require": false
          }
        ],
        "responses": {
          "200": {
            "description": "InflightAccount",
            "type": "object",
            "schema": {
              "$ref": "#/definitions/InflightAccount"
            }
          },
          "404": {
            "description": "Not Found",
            "type": "object",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "409": {
            "description": "Conflict",
            "type": "object",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "500": {
            "description": "Internal Error",
            "type": "object",
            "schema": null
          }
        }
      },
      "$ref": "#/definitions/Error",
      "get": {
        "description": "Find Inflight Account By Id",
        "operationId": "findAccountById",
        "tags": [
          "account"
        ],
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "description": "Account Id",
            "type": "string",
            "in": "path",
            "required": true
          },
          {
            "name": "X-Tenant-Id",
            "description": "Tenant Id",
            "type": "string",
            "in": "header",
            "required": false
          },
          {
            "name": "x-fast-pass",
            "description": "Bypass Auth",
            "type": "boolean",
            "in": "header",
            "require": false
          }
        ],
        "responses": {
          "200": {
            "description": "InflightAccount",
            "type": "object",
            "schema": {
              "$ref": "#/definitions/InflightAccount"
            }
          },
          "404": {
            "description": "Not Found",
            "type": "object",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "500": {
            "description": "Internal Error",
            "type": "object",
            "schema": null
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "PageResult": {
      "type": "object",
      "required": [
        "limit",
        "offset",
        "total"
      ],
      "properties": {
        "limit": {
          "type": "number"
        },
        "offset": {
          "type": "number"
        },
        "total": {
          "type": "number"
        }
      }
    },
...
    "InflightAccount": {
      "type": "object",
      "required": [
        "firstName",
        "lastName"
      ],
      "properties": {
        "id": {
          "type": "string"
        },
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "onBehalfOfTitle": {
          "type": "string"
        },
        "onBehalfOfLogoUrl": {
          "type": "string"
        },
        "avatarUrl": {
          "type": "string"
        },
        "timestamp": {
          "type": "integer"
        },
        "ttl": {
          "type": "integer"
        }
      }
    },
...
    "Error": {
      "type": "object",
      "required": [
        "errorMessage"
      ],
      "properties": {
        "errorMessage": {
          "type": "string"
        }
      }
    },
...
  },
...
  "securityDefinitions": {},
...
}